+++ src/modules/users/middlewares/index.ts
@@ -1,0 +1,1 @@
+export * from './create-user-body-validator'
